Large Refactoring
This edit refactors all the code Xingbo created into several distinct
functions, each with their own purpose-made usage. That way we can
isolate the intention of each step and modify them separately, without
breaking the whole system.

It also begins to generate the capacity for phase computations. We can
measure the state of the acrobot at each time step, but need to multiply
velocity by the inverse inertia matrix to get the phase, so we are
adding classes now which will enable us to do this in the future. These
are not fully completed, but the code verifies and should still run (it
has not been tested yet).

Since we are writing all this header code in the SUGAR_bot.h file, I
will soon refactor this into its own full library for proper phase/state
computations, and will test the code with GTest. This is to be done
ASAP, as testing the library is a high priority matter before we can get
anything else done.