Set up the VNHC code for tests
Filled out a basic VNHC for Tanh and Sin(atan2(pu,qu)), and have written
tests to guarantee that we get the correct pa value. These tests are not
completely finished, nor is the code for VNHCs. However, it is almost
there - we have our test cases, and computing the rest of the code will
be easy soon enough.