Changed function of qmax
The acrobotVNHC object now has the qmax limit code, which restricts the
input qmax to be within [0, PI]. This is to redefine the meaning of the
actuator limit; instead of it being positive or negative to represent
energy gain/loss, the limit simply specifies what the upper and lower
limits of the VNHC are (i.e. qa is restricted to be in [-qmax, qmax]).
If we want to add a distinction between energy dissipation and
injection, that will have to be done in a different way.